refactor: move deposits-aggregation from exporter to statistics and run only once per day
