+++ templates/eth1DepositsLeaderboard.html
@@ -195,6 +195,7 @@
         <div class="d-md-flex py-2 justify-content-md-between">
           <div class="heading">
             <h1 class="h4 mb-1 mb-md-0"><i class="fas fa-file-signature mr-2"></i>Validator Deposit Leaderboard</h1>
+            <div>Data on this page only gets updated once a day.</div>
           </div>
           <nav aria-label="breadcrumb">
             <ol class="breadcrumb font-size-1 mb-0" style="padding:0; background-color:transparent;">
